VideosViewControllerのビデオを消す操作を変更する。
ビデオを消す時、確認のためのAction sheetを出した。

+++ Dubbidson/VideosViewController.swift
@@ -59,23 +59,13 @@ class VideosViewController: UIViewController {
         if editing {
             let cancel = UIBarButtonItem(barButtonSystemItem: .Cancel, target: self, action: "cancelEditing")
             let space = UIBarButtonItem(barButtonSystemItem: .FlexibleSpace, target: nil, action: nil)
-            toolbar.setItems([cancel, space], animated: true)
-            //let trash = UIBarButtonItem(barButtonSystemItem: .Trash, target: self, action: "finishEditing")
-            //toolbar.setItems([cancel, space, trash], animated: true)
+            let trash = UIBarButtonItem(barButtonSystemItem: .Trash, target: self, action: "finishEditing")
+            toolbar.setItems([cancel, space, trash], animated: true)
         } else {
             let space = UIBarButtonItem(barButtonSystemItem: .FlexibleSpace, target: nil, action: nil)
             let edit = UIBarButtonItem(barButtonSystemItem: .Edit, target: self, action: "startEditing")
             toolbar.setItems([space, edit], animated: true)
         }
-        /*
-        if editing {
-            navigationItem.setLeftBarButtonItem(UIBarButtonItem(barButtonSystemItem: .Cancel, target: self, action: "cancelEditing"), animated: true)
-            navigationItem.rightBarButtonItem = nil
-        } else {
-            navigationItem.leftBarButtonItem = nil
-            navigationItem.setRightBarButtonItem(UIBarButtonItem(barButtonSystemItem: .Edit, target: self, action: "startEditing"), animated: true)
-        }
-        */
     }
 
     func startEditing() {
@@ -83,15 +73,31 @@ class VideosViewController: UIViewController {
     }
 
     func finishEditing() {
-        HUD.sharedInstance.showLoading()
-        collectionView.indexPathsForSelectedItems()?.forEach({ (indexPath) -> () in
-            //DB.sharedInstance.delete(self.videos[indexPath.row])
-            // TODO: エラー処理
-            self.videos[indexPath.row].delete()
-        })
-        HUD.sharedInstance.dismissLoading()
-        fetch()
-        setEditing(false, animated: true)
+        guard let items = collectionView.indexPathsForSelectedItems() where items.count > 0 else {
+            setEditing(false, animated: true)
+            return
+        }
+        let controller: UIAlertController = {
+            let controller = UIAlertController(title: nil, message: nil, preferredStyle: .ActionSheet)
+            controller.addAction(
+                UIAlertAction(title: "Delete \(items.count == 1 ? "Video" : "Videos")", style: .Default) { _ in
+                    HUD.sharedInstance.showLoading()
+                    items.forEach({ (indexPath) -> () in
+                        // TODO: エラー処理
+                        self.videos[indexPath.row].delete()
+                    })
+                    HUD.sharedInstance.dismissLoading()
+                    self.fetch()
+                    self.setEditing(false, animated: true)
+                })
+            controller.addAction(
+                UIAlertAction(title: "Cancel", style: .Cancel) { _ in
+                    self.fetch()
+                    self.setEditing(false, animated: true)
+                })
+            return controller
+        }()
+        presentViewController(controller, animated: true, completion: nil)
     }
 
     func cancelEditing() {
@@ -152,45 +158,12 @@ extension VideosViewController: UICollectionViewDelegate {
     func collectionView(collectionView: UICollectionView, didSelectItemAtIndexPath indexPath: NSIndexPath) {
         logger.verbose("indexPath.row: \(indexPath.row)")
         if editing {
-            if collectionView.indexPathsForSelectedItems()?.count > 0 {
-                /*
-                let cancel = UIBarButtonItem(barButtonSystemItem: .Cancel, target: self, action: "cancelEditing")
-                let space = UIBarButtonItem(barButtonSystemItem: .FlexibleSpace, target: nil, action: nil)
-                let trash = UIBarButtonItem(barButtonSystemItem: .Trash, target: self, action: "endEditing")
-                toolbar.setItems([cancel, space, trash], animated: true)
-                */
-                if toolbar.items?.count < 3 {
-                    toolbar.items?.append(UIBarButtonItem(barButtonSystemItem: .Trash, target: self, action: "finishEditing"))
-                }
-                //navigationItem.setRightBarButtonItem(UIBarButtonItem(barButtonSystemItem: .Trash, target: self, action: "finishEditing"), animated: true)
-            } else {
-                toolbar.items?.removeLast()
-                //navigationItem.rightBarButtonItem = nil
-            }
+            return
         } else {
             performSegueWithIdentifier(R.segue.watchVideo, sender: videos[indexPath.row])
         }
     }
 
-    func collectionView(collectionView: UICollectionView, didDeselectItemAtIndexPath indexPath: NSIndexPath) {
-        if editing {
-            if collectionView.indexPathsForSelectedItems()?.count > 0 {
-                /*
-                let cancel = UIBarButtonItem(barButtonSystemItem: .Cancel, target: self, action: "cancelEditing")
-                let space = UIBarButtonItem(barButtonSystemItem: .FlexibleSpace, target: nil, action: nil)
-                let trash = UIBarButtonItem(barButtonSystemItem: .Trash, target: self, action: "endEditing")
-                toolbar.setItems([cancel, space, trash], animated: true)
-                */
-                if toolbar.items?.count < 3 {
-                    toolbar.items?.append(UIBarButtonItem(barButtonSystemItem: .Trash, target: self, action: "finishEditing"))
-                }
-                //navigationItem.setRightBarButtonItem(UIBarButtonItem(barButtonSystemItem: .Trash, target: self, action: "finishEditing"), animated: true)
-            } else {
-                toolbar.items?.removeLast()
-                //navigationItem.rightBarButtonItem = nil
-            }
-        }
-    }
 }
 
 extension VideosViewController: UICollectionViewDelegateFlowLayout {
@@ -211,7 +184,6 @@ class VideoViewCell: UICollectionViewCell {
     @IBOutlet weak var nameLabel: UILabel!
     @IBOutlet weak var artistLabel: UILabel!
     //@IBOutlet weak var createdLabel: UILabel!
-    //@IBOutlet weak var checkView: UIImageView!
 
     func configure(video: Video) {
         guard let song = video.song else {
